ci: non-root container image (#437)
Make image work in a nonroot environment.

* Docker: fully-qualified image names

Consistency and compatibility with more container runtimes and builders.

* Docker: rewrite to use unprivileged NGINX

Instead of manually modifying the image to work in a non-root
environment, use the official unprivileged NGINX image. To install
packages and perform other privileged tasks, temporarily switch to root
and subsequently switch back.

Additionally, use the native envsubst templating already integrated into
the NGINX images instead of recreating it. The reason it was not
triggering before is that the entrypoint script in the base image checks
if the command is `nginx` or `nginx-debug` and if not just executes the
command directly without running any further setup.

---------

Co-authored-by: Thomas <60181704+tsk9@users.noreply.github.com>