build(ng): enable subresource integrity
Subresource integrity [1] is a feature in recent web browsers that
allows annotating scripts, stylesheets, and others with an `integrity`
attribute containing a hash of the contents.
These hashes can then be used together with CSP to ensure only trusted
scripts may be loaded.

Enabling this Angular build option only emits the `integrity` attribute
on generated <script> and <link> tags; setting a matching CSP header has
to be done in the web server configuration.

[1]: https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity